refactor: redux stores survey answers and submits them
